move AttributeTemplate to builtin_attrs
